requirements_openai.txt
openai
langchain
langchain_community
qdrant-client
sentence_transformers
pypdf
pillow
pytesseract
streamlit
python-dotenv
huggingface_hub 